Made the bus power-up have actual sprite work
with the approval of the person who originally made the bus of course.

The second file says "drifting", but it's just the turning animation, and the second to last file is the bus before Mario get's into it, the last one is when Mario is stationary and not driving the bus.

Who ever wants to impliment this, good luck! I sure didn't make it easy on ya.